piece prefabs changed into UI Item. works better
