+++ Assets/Prefabs/UI/Pieces.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: e0ca9dcec446c2840870c84aaa508c5c
-PrefabImporter:
+guid: 3c603f27fde0f1c428bf41548a8da4bf
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
