+++ Assets/Prefabs/UI/Pieces/Piece1.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14ae0574210ea724798fd760c4e1e5a2
+guid: 66f0167b8712ad94da867af2295e1f9f
 PrefabImporter:
   externalObjects: {}
   userData:
