+++ Assets/Prefabs/UI/Pieces/Piece2.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7ce7483a7ece81644a27f025006521ad
+guid: c98759e658a82dc41b7eb57ea485f07a
 PrefabImporter:
   externalObjects: {}
   userData:
