+++ Assets/Prefabs/UI/Pieces/Piece3.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 95a732718c86ac74a8e1b7ccdb63f35d
-folderAsset: yes
-DefaultImporter:
+guid: c5477a0b85c82544abfdee0cb73d8e32
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
