+++ Assets/Prefabs/UI/Pieces/Piece4.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f46179b78a22f4c40b76e529484dd064
+guid: 992c45af837abc249b4c6a85a456ec45
 PrefabImporter:
   externalObjects: {}
   userData:
