--- Assets/Scripts/Spawner/PieceSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b6de80c92ba3feb42b20c03c38d6f762
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
