+++ Assets/Scripts/UI/BasePanel.cs
@@ -4,7 +4,6 @@ using UnityEngine;
 
 public class BasePanel : MonoBehaviour
 {
-    public UIPanelType uiPanelType = UIPanelType.None;
     public virtual void Show()
     {
         gameObject.SetActive(true);
